borneo/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
mysqlclient==1.4.6
numpy==1.18.2
pandas==1.0.3
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
urllib3==1.25.8
